refactor(s3): split service/config.rs by subresource (2555 -> 1276 mod + 18 submodules)
Audit 2026-05-19 P9 continuation. S3 bucket subresource handlers split:

- bucket_inventory.rs / policy / public_access / cors / lifecycle / tagging
- encryption / replication / accelerate / website / ownership / notification
- versioning / intelligent_tiering / request_payment / logging / analytics / metrics

Each subresource family holds its put/get/delete handlers and stays under
service/config/ via pub(crate) methods called from service/mod.rs dispatch.

323 unit tests pass.
