+++ crates/fakecloud-s3/src/service/config/cors.rs
@@ -1,0 +1,101 @@
+//! `S3Service` `cors` family — extracted from service.rs by audit-2026-05-19.
+
+use super::*;
+
+impl S3Service {
+    // ---- CORS ----
+
+    pub(crate) fn put_bucket_cors(
+        &self,
+        account_id: &str,
+        req: &AwsRequest,
+        bucket: &str,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        let body_str = std::str::from_utf8(&req.body).unwrap_or("").to_string();
+
+        // Validate CORS configuration
+        let rule_count = body_str.matches("<CORSRule>").count();
+        if rule_count == 0 || rule_count > 100 {
+            return Err(AwsServiceError::aws_error(
+                StatusCode::BAD_REQUEST,
+                "MalformedXML",
+                "The XML you provided was not well-formed or did not validate against our published schema",
+            ));
+        }
+
+        // Validate HTTP methods
+        let valid_methods = ["GET", "PUT", "POST", "DELETE", "HEAD"];
+        let mut remaining = body_str.as_str();
+        while let Some(start) = remaining.find("<AllowedMethod>") {
+            let after = &remaining[start + 15..];
+            if let Some(end) = after.find("</AllowedMethod>") {
+                let method = after[..end].trim();
+                if !valid_methods.contains(&method) {
+                    return Err(AwsServiceError::aws_error(
+                        StatusCode::BAD_REQUEST,
+                        "InvalidRequest",
+                        format!(
+                            "Found unsupported HTTP method in CORS config. Unsupported method is {method}"
+                        ),
+                    ));
+                }
+                remaining = &after[end + 16..];
+            } else {
+                break;
+            }
+        }
+
+        let mut accts = self.state.write();
+        let state = accts.get_or_create(account_id);
+        let b = state
+            .buckets
+            .get_mut(bucket)
+            .ok_or_else(|| no_such_bucket(bucket))?;
+        b.cors_config = Some(body_str.clone());
+        self.store
+            .put_bucket_subresource(bucket, BucketSubresource::Cors, &body_str)
+            .map_err(crate::service::persistence_error)?;
+        Ok(empty_response(StatusCode::OK))
+    }
+
+    pub(crate) fn get_bucket_cors(
+        &self,
+        account_id: &str,
+        bucket: &str,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        let accts = self.state.read();
+        let __empty = crate::state::S3State::new(account_id, "us-east-1");
+        let state = accts.get(account_id).unwrap_or(&__empty);
+        let b = state
+            .buckets
+            .get(bucket)
+            .ok_or_else(|| no_such_bucket(bucket))?;
+        match &b.cors_config {
+            Some(config) => Ok(s3_xml(StatusCode::OK, config.clone())),
+            None => Err(AwsServiceError::aws_error_with_fields(
+                StatusCode::NOT_FOUND,
+                "NoSuchCORSConfiguration",
+                "The CORS configuration does not exist",
+                vec![("BucketName".to_string(), bucket.to_string())],
+            )),
+        }
+    }
+
+    pub(crate) fn delete_bucket_cors(
+        &self,
+        account_id: &str,
+        bucket: &str,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        let mut accts = self.state.write();
+        let state = accts.get_or_create(account_id);
+        let b = state
+            .buckets
+            .get_mut(bucket)
+            .ok_or_else(|| no_such_bucket(bucket))?;
+        b.cors_config = None;
+        self.store
+            .delete_bucket_subresource(bucket, BucketSubresource::Cors)
+            .map_err(crate::service::persistence_error)?;
+        Ok(empty_response(StatusCode::NO_CONTENT))
+    }
+}
