+++ crates/fakecloud-s3/src/service/config/intelligent_tiering.rs
@@ -1,0 +1,58 @@
+//! `S3Service` `intelligent_tiering` family — extracted from service.rs by audit-2026-05-19.
+
+use super::*;
+
+impl S3Service {
+    pub(crate) fn put_bucket_intelligent_tiering_config(
+        &self,
+        account_id: &str,
+        req: &AwsRequest,
+        bucket: &str,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        store_named_config(
+            self,
+            account_id,
+            req,
+            bucket,
+            ConfigKind::IntelligentTiering,
+        )
+    }
+
+    pub(crate) fn get_bucket_intelligent_tiering_config(
+        &self,
+        account_id: &str,
+        req: &AwsRequest,
+        bucket: &str,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        get_named_config(
+            self,
+            account_id,
+            req,
+            bucket,
+            ConfigKind::IntelligentTiering,
+        )
+    }
+
+    pub(crate) fn delete_bucket_intelligent_tiering_config(
+        &self,
+        account_id: &str,
+        req: &AwsRequest,
+        bucket: &str,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        delete_named_config(
+            self,
+            account_id,
+            req,
+            bucket,
+            ConfigKind::IntelligentTiering,
+        )
+    }
+
+    pub(crate) fn list_bucket_intelligent_tiering_configurations(
+        &self,
+        account_id: &str,
+        bucket: &str,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        list_named_config(self, account_id, bucket, ConfigKind::IntelligentTiering)
+    }
+}
