+++ crates/fakecloud-s3/src/service/config/website.rs
@@ -1,0 +1,68 @@
+//! `S3Service` `website` family — extracted from service.rs by audit-2026-05-19.
+
+use super::*;
+
+impl S3Service {
+    // ---- Website ----
+
+    pub(crate) fn put_bucket_website(
+        &self,
+        account_id: &str,
+        req: &AwsRequest,
+        bucket: &str,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        let body_str = std::str::from_utf8(&req.body).unwrap_or("").to_string();
+        let mut accts = self.state.write();
+        let state = accts.get_or_create(account_id);
+        let b = state
+            .buckets
+            .get_mut(bucket)
+            .ok_or_else(|| no_such_bucket(bucket))?;
+        b.website_config = Some(body_str.clone());
+        self.store
+            .put_bucket_subresource(bucket, BucketSubresource::Website, &body_str)
+            .map_err(crate::service::persistence_error)?;
+        Ok(empty_response(StatusCode::OK))
+    }
+
+    pub(crate) fn get_bucket_website(
+        &self,
+        account_id: &str,
+        bucket: &str,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        let accts = self.state.read();
+        let __empty = crate::state::S3State::new(account_id, "us-east-1");
+        let state = accts.get(account_id).unwrap_or(&__empty);
+        let b = state
+            .buckets
+            .get(bucket)
+            .ok_or_else(|| no_such_bucket(bucket))?;
+        match &b.website_config {
+            Some(config) => Ok(s3_xml(StatusCode::OK, config.clone())),
+            None => Err(AwsServiceError::aws_error_with_fields(
+                StatusCode::NOT_FOUND,
+                "NoSuchWebsiteConfiguration",
+                "The specified bucket does not have a website configuration",
+                vec![("BucketName".to_string(), bucket.to_string())],
+            )),
+        }
+    }
+
+    pub(crate) fn delete_bucket_website(
+        &self,
+        account_id: &str,
+        bucket: &str,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        let mut accts = self.state.write();
+        let state = accts.get_or_create(account_id);
+        let b = state
+            .buckets
+            .get_mut(bucket)
+            .ok_or_else(|| no_such_bucket(bucket))?;
+        b.website_config = None;
+        self.store
+            .delete_bucket_subresource(bucket, BucketSubresource::Website)
+            .map_err(crate::service::persistence_error)?;
+        Ok(empty_response(StatusCode::NO_CONTENT))
+    }
+}
